# gizmo-chat/gizmochat-bck/main/java/com/kvantino/gizmochat/chat/ChatResponse.java
package com.kvantino.gizmochat.chat;

import lombok.*;

import java.time.LocalDateTime;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class ChatResponse {

    private String id;
    private String name;
    private long unreadCount;
    private String lastMessage;
    private LocalDateTime lastMessageTime;
    private boolean isRecipientOnLine;
    private String senderId;
    private String receiverId;
}