# gizmo-chat/gizmochat-bck/main/java/com/kvantino/gizmochat/message/MessageRequest.java
package com.kvantino.gizmochat.message;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class MessageRequest {
    private String content;
    private String senderId;
    private String receiverId;
    private MessageType type;
    private String chatId;
}