# gizmo-chat/gizmochat-bck/main/java/com/kvantino/gizmochat/user/UserResponse.java
package com.kvantino.gizmochat.user;

import lombok.*;

import java.time.LocalDateTime;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class UserResponse {

    private String id;
    private String firstName;
    private String lastName;
    private String email;
    private LocalDateTime lastSeen;
    private boolean isOnLine;
}